feat: implement consistent card sorting for kitty cards and human hands
- Create reusable cardSorting utility with sophisticated sorting algorithm
- Apply consistent sorting to kitty cards in round complete modal
- Refactor HumanHandAnimated to use shared sorting utility
- Eliminate 130+ lines of duplicate sorting code
- Update README test count badge to reflect current 570 tests

Card sorting follows Tractor game conventions:
• Jokers first (Big → Small)
• Trump cards (rank → suit cards, Ace high)
• Non-trump cards by rotated suit order
• Maintains black-red alternation pattern
